remove ASCII-only codebase
Unicode is currently supported using \X for grapheme cluster matching,
which is only available in Perl 5.10 and newer. Support for older perls
is handled by including an ASCII-only variant of the implementation and
picking one of them at install time.

(In fact, Unicode support is only actually installed on 5.11 and newer,
even though it works on 5.10, due to an off-by-one error in the install
conditional.)

The inclusion of two separate codebases of the same module has a number
of undesirable effects on various tooling unprepared for the situation.

Additionally, the ASCII-only variant has required Perl 5.6 for a long
time. Perl 5.6 supports grapheme cluster matching using \PM, which is
less accurate than \X but suffices in practice. Therefore, shipping
a distinct ASCII-only variant of the codebase has no value.

Thus, this commit drops it.

Subsequent patches will switch Unicode support from \X to \PM, restoring
Perl 5.6 support.